pyproject.toml
[tool.poetry]
name = "button-controller"
version = "0.1.0"
description = ""
authors = ["Jacob Rogers <jacob.p.r@hotmail.com>"]
# readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
paho-mqtt = "^2.1.0"
customtkinter = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
